se transforma Login.js en index.js ya q es la entrada al sistema. Se genera register.js. Ambos front estan ok.
