+++ user_management/styles/Login.module.css
@@ -5,9 +5,7 @@
   align-items: center;
   padding-top: 10%;
   min-height: 100vh;
-  background-color: #36791E;
-  justify-content: space-between;
-  align-items: center;
+  background-color: #36791E;  
 }
 
  .login_container {
@@ -16,10 +14,6 @@
   background-color: #36791E;  
   align-items: center;
   justify-content: space-between;
-}
-
-.logo_Cice{
-  margin: auto;
 }
 
 .info_login {
@@ -47,11 +41,17 @@
 .form_login {
   width: 28rem;
   height: 29rem;
-  padding: 15px 0px;
+  padding: 1rem 1rem;
   border-radius: 20px;
   background-color: #FFFFFF;
   flex-direction: column;
   filter: drop-shadow(0px 4px 4px rgba(0, 0, 0, 0.25));
+}
+
+.register_name_surname{
+  display: flex;
+  flex-direction: row;
+  width: 95%;
 }
 
 .inputLog {
@@ -60,21 +60,22 @@
   justify-content: center;
 }
 
-.labInput {
-  margin-left: 2rem;
-  margin-top: 2rem;
+.labInput {  
+  margin: 2rem 0px 0px;
+  
 }
 
 .input {
   border-radius: 8px;
   height: 30px;
   background-color: #D9D9D9;
-  width: 80%;
+  width: 90%;
   border: none;
   outline: 1;
   margin: 0;
   padding: 0;
   margin: 0px 20px;
+
 }
 
 .inputPassword{
